requirements.txt
sexpdata==1.0.2
openai==1.54.0
rich==13.7.0
python-dotenv==1.0.0

# Optional: Anthropic Claude API for secure AI with zero-retention
# anthropic>=0.40.0

# Optional: AWS SDK for client-side encryption with KMS
# boto3>=1.34.0